vec2 int to float

+++ MVC/AbsMissile.h
@@ -6,7 +6,7 @@
 class AbsMissile : public LifetimeGameObject
 {
 public:
-	AbsMissile(Vec2<int> pos, float angle, int velocity) :
+	AbsMissile(Vec2<float> pos, float angle, int velocity) :
 		LifetimeGameObject(pos),
 		m_angle(angle),
 		m_velocity(velocity)
