+++ MVC/IGameObjectFactory.h
@@ -7,6 +7,6 @@ class AbsMissile;
 
 class IGameObjectFactory {
 public:
-	virtual AbsPlayer* createPlayer(Vec2<int> pos) = 0;
+	virtual AbsPlayer* createPlayer(Vec2<float> pos) = 0;
 	virtual AbsMissile* createMissile(float initAngle, float initVelocity) = 0;
 };
