+++ MVC/LifetimeGameObject.h
@@ -8,7 +8,7 @@ class LifetimeGameObject : public GameObject
 	using Clock = std::chrono::steady_clock;
 
 public:
-	LifetimeGameObject(Vec2<int> pos);
+	LifetimeGameObject(Vec2<float> pos);
 
 	/* Returns time of the object in milliseconds */
 	long getAge() const;
