+++ MVC/SimpleMovingStrategy.cpp
@@ -9,7 +9,7 @@ void SimpleMovingStrategy::updatePosition(AbsMissile* missile)
 	float time = 1.f - std::min(1.f, missile->getAge() / 500.f);
 	std::cout << missile->getAge() / 2000.f << std::endl;
 	missile->move({
-		(int)(missile->getVelocity() * time * cos(missile->getAngle())),
-		(int)(missile->getVelocity() * time * sin(missile->getAngle()))
+		(float)missile->getVelocity() * time * cos(missile->getAngle()),
+		(float)missile->getVelocity() * time* sin(missile->getAngle())
 		});
 }
